LayoutTests/http/tests/xmlhttprequest/event-target-expected.txt

Tests for bug 11053 - XMLHttpRequest should be an EventTarget.

fooListener: this = [object JSXMLHttpRequest]
onload: this = [object JSXMLHttpRequest]
eventListener: this = [object JSXMLHttpRequest]
eventListener2: this = [object JSXMLHttpRequest]
eventListener3: this = [object JSXMLHttpRequest]
onload: this = [object JSXMLHttpRequest]
eventListener: this = [object JSXMLHttpRequest]
eventListener2: this = [object JSXMLHttpRequest]
